+++ src/model/inline-document.ts
@@ -12,12 +12,13 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
+import * as babel from 'babel-types';
 import * as dom5 from 'dom5';
-import * as parse5 from 'parse5';
 import {ASTNode} from 'parse5';
 import * as util from 'util';
 
-import {isFakeNode} from '../html/html-document';
+import {isFakeNode, ParsedHtmlDocument} from '../html/html-document';
+import {JavaScriptDocument} from '../javascript/javascript-document';
 import * as jsdoc from '../javascript/jsdoc';
 
 import {Document, ScannedDocument} from './document';
@@ -32,6 +33,16 @@ export interface InlineDocInfo<AstNode> {
   locationOffset?: LocationOffset;
 }
 
+export type AstNodeWithLanguage = {
+  language: 'html',
+  node: dom5.Node,
+  containingDocument: ParsedHtmlDocument,
+}|{
+  language: 'js',
+  node: babel.Node,
+  containingDocument: JavaScriptDocument,
+};
+
 /**
  * Represents an inline document, usually a <script> or <style> tag in an HTML
  * document.
@@ -39,30 +50,31 @@ export interface InlineDocInfo<AstNode> {
  * @template N The AST node type
  */
 export class ScannedInlineDocument implements ScannedFeature, Resolvable {
-  type: 'html'|'js'|'css'|/* etc */ string;
+  readonly type: 'html'|'js'|'css'|/* etc */ string;
 
-  contents: string;
+  readonly contents: string;
 
   /** The location offset of this document within the containing document. */
-  locationOffset: LocationOffset;
-  attachedComment?: string;
+  readonly locationOffset: LocationOffset;
+  readonly attachedComment?: string;
 
   scannedDocument?: ScannedDocument;
 
-  sourceRange: SourceRange;
-  warnings: Warning[] = [];
+  readonly sourceRange: SourceRange;
+  readonly warnings: Warning[] = [];
 
-  astNode: dom5.Node;
+  readonly astNode: AstNodeWithLanguage;
 
   constructor(
       type: string, contents: string, locationOffset: LocationOffset,
-      attachedComment: string, sourceRange: SourceRange, ast: dom5.Node) {
+      attachedComment: string, sourceRange: SourceRange,
+      astNode: AstNodeWithLanguage) {
     this.type = type;
     this.contents = contents;
     this.locationOffset = locationOffset;
     this.attachedComment = attachedComment;
     this.sourceRange = sourceRange;
-    this.astNode = ast;
+    this.astNode = astNode;
   }
 
   resolve(document: Document): Document|undefined {
@@ -179,29 +191,13 @@ function* iterReverse<V>(arr: Array<V>): Iterable<V> {
   }
 }
 
-function isLocationInfo(loc: (parse5.LocationInfo|parse5.ElementLocationInfo)):
-    loc is parse5.LocationInfo {
-  return 'line' in loc;
-}
-
-export function getLocationOffsetOfStartOfTextContent(node: ASTNode):
-    LocationOffset {
-  const childNodes = node.childNodes || [];
-  const firstChildNodeWithLocation = childNodes.find((n) => !!n.__location);
-  const bestLocation = firstChildNodeWithLocation ?
-      firstChildNodeWithLocation.__location :
-      node.__location;
-  if (!bestLocation) {
+export function getLocationOffsetOfStartOfTextContent(
+    node: ASTNode, parsedDocument: ParsedHtmlDocument): LocationOffset {
+  const sourceRange = parsedDocument.sourceRangeForStartTag(node);
+  if (!sourceRange) {
     throw new Error(
         `Couldn't extract a location offset from HTML node: ` +
         `${util.inspect(node)}`);
   }
-  if (isLocationInfo(bestLocation)) {
-    return {line: bestLocation.line - 1, col: bestLocation.col - 1};
-  } else {
-    return {
-      line: bestLocation.startTag.line - 1,
-      col: bestLocation.startTag.endOffset - 1,
-    };
-  }
+  return {line: sourceRange.end.line, col: sourceRange.end.column};
 }